SemVer 2.0.0 compliant version for nightly builds
[SemVer 2.0.0](https://semver.org) (used for NuGet packages):
$9: A pre-release version MAY be denoted by appending a hyphen and a series of dot separated identifiers immediately following the patch version.
§10: Build metadata MAY be denoted by appending a plus sign and a series of dot separated identifiers immediately following the patch or pre-release version.

Adding date to build metadata would prevent packages from being pushed to repos and downloaded by NuGet clients. Thus it must to be on the left side of the plus sign in order to give the package precedence. It also makes sense to use the date as version for a nightly build.